+++ src/input.rs
@@ -22,14 +22,14 @@ pub use config::{
     CheckboxChildrenConfig, CheckboxConfig, ChildrenPosition, ComparisonConfig, Essential,
     FileConfig, Float64Config, GroupConfig, HostNetworkGroupConfig, InputConfig, NicConfig,
     PasswordConfig, PercentageConfig, RadioConfig, SelectMultipleConfig, SelectSingleConfig,
-    TagConfig, TextConfig, Unsigned32Config, VecSelectConfig,
+    TagConfig, TextConfig, Unsigned32Config, Unsigned8Config, VecSelectConfig,
 };
 pub use host_network::Kind as HostNetworkKind;
 pub use host_network::Model as HostNetworkHtml;
 pub use item::{
     CheckboxItem, ComparisonItem, FileItem, Float64Item, GroupItem, HostNetworkGroupItem,
     InputItem, NicItem, PasswordItem, PercentageItem, RadioItem, SelectMultipleItem,
-    SelectSingleItem, TagItem, TextItem, Unsigned32Item, VecSelectItem,
+    SelectSingleItem, TagItem, TextItem, Unsigned32Item, Unsigned8Item, VecSelectItem,
 };
 use num_bigint::BigUint;
 use num_traits::ToPrimitive;
@@ -529,6 +529,7 @@ fn default_items(confs: &[Rc<InputConfig>], level: usize) -> Vec<Rc<RefCell<Inpu
                 }
                 InputConfig::Tag(_) => InputItem::Tag(TagItem::default()),
                 InputConfig::Unsigned32(_) => InputItem::Unsigned32(Unsigned32Item::default()),
+                InputConfig::Unsigned8(_) => InputItem::Unsigned8(Unsigned8Item::default()),
                 InputConfig::Float64(_) => InputItem::Float64(Float64Item::default()),
                 InputConfig::Percentage(_) => InputItem::Percentage(PercentageItem::default()),
                 InputConfig::Nic(_) => InputItem::Nic(NicItem::default()),
@@ -604,6 +605,7 @@ fn item_preset(conf: &Rc<InputConfig>) -> InputItem {
             }),
         )),
         InputConfig::Unsigned32(conf) => InputItem::Unsigned32(Unsigned32Item::new(conf.preset)),
+        InputConfig::Unsigned8(conf) => InputItem::Unsigned8(Unsigned8Item::new(conf.preset)),
         InputConfig::Float64(conf) => InputItem::Float64(Float64Item::new(conf.preset)),
         InputConfig::Percentage(conf) => InputItem::Percentage(PercentageItem::new(conf.preset)),
         InputConfig::Comparison(_) => InputItem::Comparison(ComparisonItem::new(None)),