+++ src/input/item.rs
@@ -341,6 +341,50 @@ impl Unsigned32Item {
 }
 
 #[derive(Clone, PartialEq, Default)]
+pub struct Unsigned8Item {
+    value: Option<u8>,
+}
+
+impl Deref for Unsigned8Item {
+    type Target = Option<u8>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.value
+    }
+}
+
+impl DerefMut for Unsigned8Item {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.value
+    }
+}
+
+impl Unsigned8Item {
+    #[must_use]
+    pub fn new(value: Option<u8>) -> Self {
+        Self { value }
+    }
+
+    pub fn set(&mut self, value: u8) {
+        self.value = Some(value);
+    }
+
+    #[must_use]
+    pub fn is_empty(&self) -> bool {
+        self.value.is_none()
+    }
+
+    #[must_use]
+    pub fn into_inner(&self) -> Option<u8> {
+        self.value
+    }
+
+    pub fn clear(&mut self) {
+        *self = Self::new(None);
+    }
+}
+
+#[derive(Clone, PartialEq, Default)]
 pub struct Float64Item {
     value: Option<f64>,
 }
@@ -764,6 +808,7 @@ pub enum InputItem {
     SelectMultiple(SelectMultipleItem),
     Tag(TagItem),
     Unsigned32(Unsigned32Item),
+    Unsigned8(Unsigned8Item),
     Float64(Float64Item),
     Percentage(PercentageItem),
     Nic(NicItem),
@@ -785,6 +830,7 @@ impl InputItem {
             InputItem::SelectMultiple(selected_list) => selected_list.clear(),
             InputItem::Tag(tag) => tag.clear(),
             InputItem::Unsigned32(value) => value.clear(),
+            InputItem::Unsigned8(value) => value.clear(),
             InputItem::Float64(value) => value.clear(),
             InputItem::Percentage(value) => value.clear(),
             InputItem::Nic(nics) => nics.clear(),
@@ -807,6 +853,7 @@ impl InputItem {
             InputItem::SelectMultiple(selected_list) => selected_list.is_empty(),
             InputItem::Tag(tag) => tag.is_empty(),
             InputItem::Unsigned32(value) => value.is_empty(),
+            InputItem::Unsigned8(value) => value.is_empty(),
             InputItem::Float64(value) => value.is_empty(),
             InputItem::Percentage(value) => value.is_empty(),
             InputItem::Nic(nics) => nics.is_empty(),
@@ -830,6 +877,7 @@ impl InputItem {
             InputConfig::SelectMultiple(_) => Self::SelectMultiple(SelectMultipleItem::default()),
             InputConfig::Tag(_) => Self::Tag(TagItem::default()),
             InputConfig::Unsigned32(_) => Self::Unsigned32(Unsigned32Item::default()),
+            InputConfig::Unsigned8(_) => Self::Unsigned8(Unsigned8Item::default()),
             InputConfig::Float64(_) => Self::Float64(Float64Item::default()),
             InputConfig::Percentage(_) => Self::Percentage(PercentageItem::default()),
             InputConfig::Nic(_) => Self::Nic(NicItem::default()),
@@ -888,6 +936,7 @@ impl From<&Column> for InputItem {
                 delete: None,
             })),
             Column::Unsigned32(value) => Self::Unsigned32(Unsigned32Item::new(value.value)),
+            Column::Unsigned8(value) => Self::Unsigned8(Unsigned8Item::new(value.value)),
             Column::Float64(value) => Self::Float64(Float64Item::new(value.value)),
             Column::Percentage(value) => Self::Percentage(PercentageItem::new(value.value)),
             Column::Nic(nics) => Self::Nic(NicItem::new(nics.nics.clone())),
@@ -921,6 +970,7 @@ impl From<&Column> for InputItem {
                             | Column::SelectSingle(..)
                             | Column::SelectMultiple(..)
                             | Column::Unsigned32(..)
+                            | Column::Unsigned8(..)
                             | Column::Float64(..)
                             | Column::Percentage(..)
                             | Column::Comparison(..)